refactor(mac): migrate more gpui cocoa bindings to objc2
Migrate additional macOS platform code from deprecated cocoa APIs to objc2-app-kit and objc2-foundation, including dialog, display, dock, events, and platform lifecycle wiring.

Also fix NSScreenNumber extraction in display handling to preserve the downstream BananaTray tray-anchor path while reducing deprecation warnings.

+++ crates/gpui/src/platform/mac/window.rs
@@ -2685,7 +2685,7 @@ unsafe fn remove_layer_background(layer: id) {
 
                 let all_indices = NSRange {
                     location: 0,
-                    length: count,
+                    length: count as usize,
                 };
                 let indices: id = msg_send![class!(NSMutableIndexSet), indexSet];
                 let _: () = msg_send![indices, addIndexesInRange: all_indices];
